Fix monitoring calibration, deep-model scoring, and add a retrain cooldown
Found by actually running the loop rather than by reading the code.

Deep model scored nothing (silent)
- predict_panel needs sequence_length contiguous rows per entity, but the
  validation split IS exactly 168 hours and sequence_length is 168, so it
  produced ZERO sequences. The candidate was logged with no usable validation
  predictions and early stopping silently fell back to the training loss.
- Splits now draw sequence context from preceding data, matching how serving
  works. Validation sequences went 0 -> 6,720 (full coverage), test -> 13,480.

Monitoring fired on clean data (two causes)
- mae_threshold of 4.0 is meaningless when mean demand is ~161/cluster-hour and
  the model's own holdout MAE is 24.6; rolling MAE of 23.8 was BETTER than the
  holdout yet counted as a breach. Ceiling is now the seasonal-naive baseline
  (45), with the scale-free degradation ratio as the primary rule.
- Calendar features are deterministic functions of the timestamp, so a 24h
  window differs from a 9-week reference by construction. They drove most of an
  apparent 77% drift share on undrifted data and are now excluded.
- The drift remaining after both fixes is real: late-June demand is 1.22x the
  training-window average at 26.9C vs 16.8C. Verified against the data.

Retrain storm
- Drift persists until a new model serves, so an hourly monitor queued a fresh
  retrain every hour (3 had already stacked up). Added a 6h cooldown that fails
  open and never extends itself.

MLflow signature fragility
- Inferred signatures declared integer columns, so features read back out of
  the JSONB features table (all float after the JSON round-trip) were rejected.
  MLflow warns about exactly this at log time. Signatures are now all-double.

Verified in Airflow: monitor_dag -> "Branch into trigger_retrain" -> retrain_dag
created and running, with no human in the loop.

110 tests passing, ruff clean.

+++ src/models/pyfunc_wrapper.py
@@ -65,7 +65,10 @@ class DemandModelWrapper(mlflow.pyfunc.PythonModel):
                 f"Input is missing {len(missing)} feature column(s) required by this "
                 f"model: {missing[:10]}"
             )
-        matrix = model_input[self.features]
+        # Everything downstream is numeric, so coerce to float. This keeps a
+        # caller that supplies ints (or strings from a JSON payload) working
+        # rather than failing deep inside the estimator.
+        matrix = model_input[self.features].astype("float64")
         raw = self._booster.predict(matrix)
         return np.clip(np.asarray(raw, dtype=float), 0.0, None)
